cloud_function_files/requirements.txt
google-cloud-monitoring
google-cloud-bigquery
datetime
argparse